+++ packages/element/src/binding.ts
@@ -226,10 +226,7 @@ const getOriginalBindingIfStillCloseOfLinearElementEdge = (
       : linearElement.endBinding?.elementId;
   if (elementId) {
     const element = elementsMap.get(elementId);
-    if (
-      isBindableElement(element) &&
-      bindingBorderTest(element, coors, elementsMap, zoom)
-    ) {
+    if (isBindableElement(element) && bindingBorderTest(element, coors, zoom)) {
       return element;
     }
   }
@@ -559,7 +556,6 @@ export const getHoveredElementForBinding = (
         bindingBorderTest(
           element,
           pointerCoords,
-          elementsMap,
           zoom,
           (fullShape ||
             !isBindingFallthroughEnabled(
@@ -592,7 +588,7 @@ export const getHoveredElementForBinding = (
 
     // Prefer the shape with the border being tested (if any)
     const borderTestElements = candidateElements.filter((element) =>
-      bindingBorderTest(element, pointerCoords, elementsMap, zoom, false),
+      bindingBorderTest(element, pointerCoords, zoom, false),
     );
     if (borderTestElements.length === 1) {
       return borderTestElements[0];
@@ -613,7 +609,6 @@ export const getHoveredElementForBinding = (
       bindingBorderTest(
         element,
         pointerCoords,
-        elementsMap,
         zoom,
         // disable fullshape snapping for frame elements so we
         // can bind to frame children
@@ -1545,7 +1540,6 @@ const newBoundElements = (
 export const bindingBorderTest = (
   element: NonDeleted<ExcalidrawBindableElement>,
   { x, y }: { x: number; y: number },
-  elementsMap: NonDeletedSceneElementsMap,
   zoom?: AppState["zoom"],
   fullShape?: boolean,
 ): boolean => {
