+++ packages/element/src/bounds.ts
@@ -9,29 +9,22 @@ import {
 
 import {
   degreesToRadians,
-  lineSegment,
   pointDistance,
   pointFrom,
   pointFromArray,
   pointRotateRads,
 } from "@excalidraw/math";
 
-import { getCurvePathOps } from "@excalidraw/utils/shape";
-
-import { pointsOnBezierCurves } from "points-on-curve";
-
 import type {
-  Curve,
   Degrees,
   GlobalPoint,
-  LineSegment,
   LocalPoint,
   Radians,
 } from "@excalidraw/math";
 
 import type { AppState } from "@excalidraw/excalidraw/types";
 
-import type { Mutable } from "@excalidraw/common/utility-types";
+import { getCurvePathOps } from "./utils";
 
 import { generateRoughOptions } from "./Shape";
 import { ShapeCache } from "./ShapeCache";
@@ -45,13 +38,6 @@ import {
   isTextElement,
 } from "./typeChecks";
 
-import { getElementShape } from "./shapes";
-
-import {
-  deconstructDiamondElement,
-  deconstructRectanguloidElement,
-} from "./utils";
-
 import type { Drawable, Op } from "roughjs/bin/core";
 import type { Point as RoughPoint } from "roughjs/bin/geometry";
 import type {
@@ -59,10 +45,8 @@ import type {
   ElementsMap,
   ElementsMapOrArray,
   ExcalidrawElement,
-  ExcalidrawEllipseElement,
   ExcalidrawFreeDrawElement,
   ExcalidrawLinearElement,
-  ExcalidrawRectanguloidElement,
   ExcalidrawTextElementWithContainer,
   NonDeleted,
 } from "./types";
@@ -265,199 +249,6 @@ export const getElementAbsoluteCoords = (
     element.x + element.width / 2,
     element.y + element.height / 2,
   ];
-};
-
-/*
- * for a given element, `getElementLineSegments` returns line segments
- * that can be used for visual collision detection (useful for frames)
- * as opposed to bounding box collision detection
- */
-/**
- * Given an element, return the line segments that make up the element.
- *
- * Uses helpers from /math
- */
-export const getElementLineSegments = (
-  element: ExcalidrawElement,
-  elementsMap: ElementsMap,
-): LineSegment<GlobalPoint>[] => {
-  const shape = getElementShape(element, elementsMap);
-  const [x1, y1, x2, y2, cx, cy] = getElementAbsoluteCoords(
-    element,
-    elementsMap,
-  );
-  const center = pointFrom<GlobalPoint>(cx, cy);
-
-  if (shape.type === "polycurve") {
-    const curves = shape.data;
-    const points = curves
-      .map((curve) => pointsOnBezierCurves(curve, 10))
-      .flat();
-    let i = 0;
-    const segments: LineSegment<GlobalPoint>[] = [];
-    while (i < points.length - 1) {
-      segments.push(
-        lineSegment(
-          pointFrom(points[i][0], points[i][1]),
-          pointFrom(points[i + 1][0], points[i + 1][1]),
-        ),
-      );
-      i++;
-    }
-
-    return segments;
-  } else if (shape.type === "polyline") {
-    return shape.data as LineSegment<GlobalPoint>[];
-  } else if (_isRectanguloidElement(element)) {
-    const [sides, corners] = deconstructRectanguloidElement(element);
-    const cornerSegments: LineSegment<GlobalPoint>[] = corners
-      .map((corner) => getSegmentsOnCurve(corner, center, element.angle))
-      .flat();
-    const rotatedSides = getRotatedSides(sides, center, element.angle);
-    return [...rotatedSides, ...cornerSegments];
-  } else if (element.type === "diamond") {
-    const [sides, corners] = deconstructDiamondElement(element);
-    const cornerSegments = corners
-      .map((corner) => getSegmentsOnCurve(corner, center, element.angle))
-      .flat();
-    const rotatedSides = getRotatedSides(sides, center, element.angle);
-
-    return [...rotatedSides, ...cornerSegments];
-  } else if (shape.type === "polygon") {
-    if (isTextElement(element)) {
-      const container = getContainerElement(element, elementsMap);
-      if (container && isLinearElement(container)) {
-        const segments: LineSegment<GlobalPoint>[] = [
-          lineSegment(pointFrom(x1, y1), pointFrom(x2, y1)),
-          lineSegment(pointFrom(x2, y1), pointFrom(x2, y2)),
-          lineSegment(pointFrom(x2, y2), pointFrom(x1, y2)),
-          lineSegment(pointFrom(x1, y2), pointFrom(x1, y1)),
-        ];
-        return segments;
-      }
-    }
-
-    const points = shape.data as GlobalPoint[];
-    const segments: LineSegment<GlobalPoint>[] = [];
-    for (let i = 0; i < points.length - 1; i++) {
-      segments.push(lineSegment(points[i], points[i + 1]));
-    }
-    return segments;
-  } else if (shape.type === "ellipse") {
-    return getSegmentsOnEllipse(element as ExcalidrawEllipseElement);
-  }
-
-  const [nw, ne, sw, se, , , w, e] = (
-    [
-      [x1, y1],
-      [x2, y1],
-      [x1, y2],
-      [x2, y2],
-      [cx, y1],
-      [cx, y2],
-      [x1, cy],
-      [x2, cy],
-    ] as GlobalPoint[]
-  ).map((point) => pointRotateRads(point, center, element.angle));
-
-  return [
-    lineSegment(nw, ne),
-    lineSegment(sw, se),
-    lineSegment(nw, sw),
-    lineSegment(ne, se),
-    lineSegment(nw, e),
-    lineSegment(sw, e),
-    lineSegment(ne, w),
-    lineSegment(se, w),
-  ];
-};
-
-const _isRectanguloidElement = (
-  element: ExcalidrawElement,
-): element is ExcalidrawRectanguloidElement => {
-  return (
-    element != null &&
-    (element.type === "rectangle" ||
-      element.type === "image" ||
-      element.type === "iframe" ||
-      element.type === "embeddable" ||
-      element.type === "frame" ||
-      element.type === "magicframe" ||
-      (element.type === "text" && !element.containerId))
-  );
-};
-
-const getRotatedSides = (
-  sides: LineSegment<GlobalPoint>[],
-  center: GlobalPoint,
-  angle: Radians,
-) => {
-  return sides.map((side) => {
-    return lineSegment(
-      pointRotateRads<GlobalPoint>(side[0], center, angle),
-      pointRotateRads<GlobalPoint>(side[1], center, angle),
-    );
-  });
-};
-
-const getSegmentsOnCurve = (
-  curve: Curve<GlobalPoint>,
-  center: GlobalPoint,
-  angle: Radians,
-): LineSegment<GlobalPoint>[] => {
-  const points = pointsOnBezierCurves(curve, 10);
-  let i = 0;
-  const segments: LineSegment<GlobalPoint>[] = [];
-  while (i < points.length - 1) {
-    segments.push(
-      lineSegment(
-        pointRotateRads<GlobalPoint>(
-          pointFrom(points[i][0], points[i][1]),
-          center,
-          angle,
-        ),
-        pointRotateRads<GlobalPoint>(
-          pointFrom(points[i + 1][0], points[i + 1][1]),
-          center,
-          angle,
-        ),
-      ),
-    );
-    i++;
-  }
-
-  return segments;
-};
-
-const getSegmentsOnEllipse = (
-  ellipse: ExcalidrawEllipseElement,
-): LineSegment<GlobalPoint>[] => {
-  const center = pointFrom<GlobalPoint>(
-    ellipse.x + ellipse.width / 2,
-    ellipse.y + ellipse.height / 2,
-  );
-
-  const a = ellipse.width / 2;
-  const b = ellipse.height / 2;
-
-  const segments: LineSegment<GlobalPoint>[] = [];
-  const points: GlobalPoint[] = [];
-  const n = 90;
-  const deltaT = (Math.PI * 2) / n;
-
-  for (let i = 0; i < n; i++) {
-    const t = i * deltaT;
-    const x = center[0] + a * Math.cos(t);
-    const y = center[1] + b * Math.sin(t);
-    points.push(pointRotateRads(pointFrom(x, y), center, ellipse.angle));
-  }
-
-  for (let i = 0; i < points.length - 1; i++) {
-    segments.push(lineSegment(points[i], points[i + 1]));
-  }
-
-  segments.push(lineSegment(points[points.length - 1], points[0]));
-  return segments;
 };
 
 /**
@@ -868,7 +659,7 @@ const generateLinearElementShape = (
   })();
 
   return generator[method](
-    element.points as Mutable<LocalPoint>[] as RoughPoint[],
+    element.points as LocalPoint[] as RoughPoint[],
     options,
   );
 };
