+++ packages/element/src/linearElementEditor.ts
@@ -9,8 +9,6 @@ import {
   vectorFromPoint,
 } from "@excalidraw/math";
 
-import { getCurvePathOps } from "@excalidraw/utils/shape";
-
 import {
   DRAGGING_THRESHOLD,
   KEYS,
@@ -20,7 +18,7 @@ import {
   tupleToCoors,
 } from "@excalidraw/common";
 
-import type { Store } from "@excalidraw/element";
+import { getCurvePathOps, type Store } from "@excalidraw/element";
 
 import type { Radians } from "@excalidraw/math";
