+++ packages/excalidraw/components/App.tsx
@@ -128,7 +128,7 @@ import {
   refreshTextDimensions,
   deepCopyElement,
   duplicateElements,
-  isPointInShape,
+  isPointInElement,
   hasBoundTextElement,
   isArrowElement,
   isBindingElement,
@@ -5159,7 +5159,7 @@ class App extends React.Component<AppProps, AppState> {
     ) {
       // if hitting the bounding box, return early
       // but if not, we should check for other cases as well (e.g. frame name)
-      if (isPointInShape(pointFrom(x, y), element)) {
+      if (isPointInElement(pointFrom(x, y), element)) {
         return true;
       }
     }
